third_party/blink/web_tests/platform/mac/virtual/no-auto-wpt-origin-isolation/external/wpt/html/browsers/origin/origin-keyed-agent-clusters/1-iframe/parent-yes-child-no-port.sub.https-expected.txt

This is a testharness.js-based test.
PASS messageerror event must occur
FAIL setting document.domain must not give sync access assert_throws_dom: function "() => {
        frameWindow.document;
      }" did not throw
FAIL parent: originAgentCluster must equal true assert_equals: expected true but got false
PASS child: originAgentCluster must equal false
Harness: the test ran to completion.
